refactorizacion de el directorio de interfaces
