+++ src/app/(Recipes pages)/functions/GetAllRecipesOneCountry.ts
@@ -1,5 +1,5 @@
-import { type Recipes } from '@/interfaces/Recipes.interface'
 import { baseUrl } from '@/lib/utils'
+import { type Recipes } from '@interfaces/Recipes.interface'
 
 /**
  * @description Obtiene una lista de recetas de un país específico desde el servidor.
@@ -8,17 +8,17 @@ import { baseUrl } from '@/lib/utils'
  */
 
 export async function GetAllRecipesOneCountry(id: string) {
-	try {
-		const response = await fetch(
-			`${baseUrl}/recipes/filter/countrys?countryId=${id}`,
-		)
-		if (!response.ok) {
-			throw new Error('Error al obtener todas las recetas de 1 pais')
-		}
-		const data: Recipes[] = await response.json()
+  try {
+    const response = await fetch(
+      `${baseUrl}/recipes/filter/countrys?countryId=${id}`
+    )
+    if (!response.ok) {
+      throw new Error('Error al obtener todas las recetas de 1 pais')
+    }
+    const data: Recipes[] = await response.json()
 
-		return data
-	} catch (error) {
-		console.error(error)
-	}
+    return data
+  } catch (error) {
+    console.error(error)
+  }
 }
