+++ src/app/(Recipes pages)/paises/[id]/page.tsx
@@ -3,7 +3,7 @@ import CardRecipe from '@/components/shared/CardRecipe'
 import DynamicTitle from '@/components/shared/DynamicTitle'
 import NotFound from '@/components/shared/NotFound'
 import ShowOneRecipe from '@/components/shared/ShowOneRecipe'
-import { type ParamProps } from '@/interfaces/ParamProps.interface'
+import { type ParamProps } from '@interfaces/ParamProps.interface'
 import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import { Suspense } from 'react'
