+++ src/app/(Recipes pages)/recetas/components/oneRecipe/CheckBox.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { type ChildrenProps } from '@/interfaces/ChildrenProps.interface'
+import { type ChildrenProps } from '@interfaces/ChildrenProps.interface'
 import { Checkbox } from '@nextui-org/react'
 
 const CheckBoxComponent = ({ children }: ChildrenProps) => {
