+++ src/app/(Recipes pages)/recetas/components/oneRecipe/Chip.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { type ChildrenProps } from '@/interfaces/ChildrenProps.interface'
+import { type ChildrenProps } from '@interfaces/ChildrenProps.interface'
 import { Chip } from '@nextui-org/react'
 
 const ChipComponent = ({ children }: ChildrenProps) => {
