+++ src/interfaces/CardCountryProps.interface.ts
