+++ src/interfaces/ParamProps.interface.ts
