+++ src/interfaces/Recipes.interface.ts
@@ -1,0 +1,17 @@
+import { type Categories } from '@interfaces/Categories.interface'
+import { type Country } from '@interfaces/Cuntry.interface'
+
+export interface Recipes {
+  _id: string
+  name: string
+  description: string
+  ingredients: string[]
+  steps: string[]
+  category: Categories
+  duration: number
+  portions: number
+  country?: Country
+  image: string
+  createdAt: string
+  updatedAt: string
+}
